random agent layout functionality

+++ herding/opencl/rand.h
@@ -1,5 +1,5 @@
 
-float rand(unsigned int *seed, int range)
+float rand(unsigned long *seed, int range)
 {
     *seed = ((*seed * 214013L + 2531011L) >> 16);
     return *seed % range;
